+++ sympy/core/__init__.py
@@ -8,7 +8,8 @@ from .singleton import S
 from .expr import Expr, AtomicExpr
 from .symbol import Symbol, Wild, Dummy, symbols, var
 from .numbers import Number, Float, Rational, Integer, NumberSymbol, \
-    RealNumber, igcd, ilcm, seterr, E, I, nan, oo, pi, zoo
+    RealNumber, igcd, ilcm, seterr, E, I, nan, oo, pi, zoo, \
+    AlgebraicNumber
 from .power import Pow, integer_nthroot
 from .mul import Mul, prod
 from .add import Add
